Defined size as unsigned int. Size of size_t (= unsigned long) and unsigned int differs on architectures with LP64 data model.
git-svn-id: svn://nex.scrapping.cc/shell-fm/trunk@38 7df44517-d413-0410-91cf-82ca28b36b55

+++ autoban.c
@@ -20,7 +20,7 @@ int banned(const char * artist) {
 	FILE * fd;
 	signed match = 0;
 	char * line = NULL;
-	size_t size = 0;
+	unsigned int size = 0;
 	
 	if(!artist)
 		return 0;
